Config: Make POLL_AWAIT_TIME_MS_DEFAULT long as the config returns a long

+++ src/main/java/com/mongodb/kafka/connect/source/MongoSourceConfig.java
@@ -265,7 +265,7 @@ public class MongoSourceConfig extends AbstractConfig {
   public static final String POLL_AWAIT_TIME_MS_CONFIG = "poll.await.time.ms";
   private static final String POLL_AWAIT_TIME_MS_DOC =
       "The amount of time to wait before checking for new results on the change stream";
-  private static final int POLL_AWAIT_TIME_MS_DEFAULT = 5000;
+  private static final long POLL_AWAIT_TIME_MS_DEFAULT = 5000;
   private static final String POLL_AWAIT_TIME_MS_DISPLAY = "Poll await time (ms)";
 
   public static final String DATABASE_CONFIG = "database";
